Refactored bodies/fixtures creation.

+++ core/src/com/aesophor/medievania/util/box2d/BodyBuilder.java
@@ -175,7 +175,7 @@ public class BodyBuilder {
      * @param maskBits mask bits.
      * @return BodyBuilder instance.
      */
-    public BodyBuilder maskBits(int maskBits) {
+    public BodyBuilder maskBits(short maskBits) {
         fdef.filter.maskBits = (short) maskBits;
         return this;
     }
